Removido e renomeado alguns arquivos
